% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.query_series_by_region}
\alias{.query_series_by_region}
\title{Regional Soil Series Data Download}
\usage{
.query_series_by_region(remDr, x, start_year = NULL, end_year = NULL)
}
\arguments{
\item{remDr}{RSelenium Client}

\item{x}{A value}

\item{start_year}{Optional: Start of target "origin" year interval}

\item{end_year}{Optional: End of target "origin" year interval}
}
\value{
Path to a Zip File containing series from the selected region.
}
\description{
Download ZIP files containing multiple (>1000) Official Series Descriptions (OSDs) into local Downloads folder. Files matching "osddwn.*zip$" are moved to the package raw directory. The function will wait up to 2 minutes for the server to complete the ZIP file result, but will complete as soon as the result is detected.
}
\details{
Need to maintain a single session throughout a sequence of horrid ASPX form steps. Individual sessions maintained via __VIEWSTATE, __EVENTVALIDATION and other "secret" fields.
}
\author{
Andrew G. Brown
}
