% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{refresh_registry}
\alias{refresh_registry}
\title{Refresh the repository/package data by automating regional queries to soil series server}
\usage{
refresh_registry()
}
\value{
Text files written to alphabetical folders containing raw Official Series Description text. Derivative calculations output to X. Etc.
}
\description{
Refresh the repository/package data by automating regional queries to soil series server
}
