% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{refresh_registry}
\alias{refresh_registry}
\title{Automatic update by regional queries to NASIS soil series server}
\usage{
refresh_registry(
  test = FALSE,
  moID = c(Alaska = 36871, `North Central` = 117, Northeast = 153, Northwest = 134,
    `South Central` = 127, Southeast = 122, Southwest = 113, `Special Projects` = 44372),
  port = 4567L
)
}
\arguments{
\item{test}{Default: \code{FALSE}; run on a pair of small regions (MO 12, 13)}

\item{moID}{Region ID codes; see default argument value in function definition for details}

\item{port}{Passed to \code{\link[RSelenium:rsDriver]{RSelenium::rsDriver()}}. Default: \code{4567L}.}
}
\value{
\code{0} if function completes.
}
\description{
Text files are written to alphabetical (first letter) folders containing raw Official Series Descriptions (OSDs). This method is for use in automatic pipeline (e.g. a GitHub action) to regularly replicate changes that occur across the entire set of series for commit.

There is an assumption that the files are downloaded to a Linux-like default \code{"Downloads"} folder \code{"~/Downloads"} or \code{"/home/user/Downloads"} which are standard on \code{"ubuntu-latest"} where these actions are typically run. The files matching the path \code{"osddwn.*zip$"} get moved to the repository "raw" folder.
}
\details{
Thank goodness we don't need to read .doc files... the server delivers .txt masquerading as .doc. Queries that error are re-tried after splitting in half (established year before or after 1980).
}
\examples{

## note: takes several minutes to run and downloads ~100MB .ZIP data
# refresh_registry()

}
\author{
Andrew G. Brown
}
