% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{refresh_registry}
\alias{refresh_registry}
\title{Refresh the repository/package data by automating regional queries to soil series server}
\usage{
refresh_registry()
}
\value{
Text files written to alphabetical folders containing raw Official Series Description (OSD) text. For use in automatic pipeline to regularly commit changes across the entire set of OSDs.
}
\description{
Thank goodness... I forgot we actually don't need to read .doc files... the server delivers .txt masquerading as .doc.
}
